Add files via upload
The project records a GET request each time an email is opened, which is used to store information about the email, including the time it was opened and the number of times it was accessed. This information is stored in a MySQL database and can be used to track email opens and other events.